+++ src/main/java/domain/game/Referee.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 public class Referee {
     private static final int BUST_THRESHOLD = 21;
