+++ src/main/java/domain/game/Result.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 public enum Result {
     WIN, LOSE, TIE
