+++ src/test/java/domain/bet/BettingMoneyTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.bet;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
