+++ src/test/java/domain/card/CardBundleTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
