+++ src/test/java/domain/game/RefereeTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
